# openapi.yaml
openapi: 3.0.3
info:
  title: FitReg API
  description: |
    REST API for FitReg, a fitness coaching platform where coaches assign and manage
    workouts for their students.

    **Authentication:** All endpoints except `/health` and `POST /api/auth/google`
    require a JWT Bearer token obtained via Google OAuth.
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/workouts:
    get:
      tags: [workouts]
      summary: List own workouts
      responses:
        "200":
          description: List of workouts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workout"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [workouts]
      summary: Create workout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkoutRequest"
      responses:
        "201":
          description: Created workout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workout"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/coach/students:
    get:
      tags: [coach-students]
      summary: List active students
      responses:
        "200":
          description: Students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CoachStudentInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/coach/assigned-workouts:
    get:
      tags: [coach-assigned-workouts]
      summary: List all assigned workouts (coach view)
      responses:
        "200":
          description: Assigned workouts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssignedWorkout"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [coach-assigned-workouts]
      summary: Create assigned workout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssignedWorkoutRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignedWorkout"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/my-assigned-workouts:
    get:
      tags: [student-assigned-workouts]
      summary: Get own assigned workouts
      responses:
        "200":
          description: Assigned workouts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssignedWorkout"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/admin/users:
    get:
      tags: [admin]
      summary: List all users
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    WorkoutSegment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        order_index:
          type: integer
        segment_type:
          type: string
          enum: [simple, interval]
        repetitions:
          type: integer
        value:
          type: number
          format: float
        unit:
          type: string
          example: km
        intensity:
          type: string
          example: easy
        work_value:
          type: number
          format: float
        work_unit:
          type: string
        work_intensity:
          type: string
        rest_value:
          type: number
          format: float
        rest_unit:
          type: string
        rest_intensity:
          type: string
    SegmentRequest:
      type: object
      required: [segment_type]
      properties:
        segment_type:
          type: string
          enum: [simple, interval]
        repetitions:
          type: integer
          default: 1
        value:
          type: number
          format: float
        unit:
          type: string
        intensity:
          type: string
        work_value:
          type: number
          format: float
        work_unit:
          type: string
        work_intensity:
          type: string
        rest_value:
          type: number
          format: float
        rest_unit:
          type: string
        rest_intensity:
          type: string
    Workout:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        date:
          type: string
          format: date
        distance_km:
          type: number
          format: float
        duration_seconds:
          type: integer
        avg_pace:
          type: string
        calories:
          type: integer
        avg_heart_rate:
          type: integer
        feeling:
          type: integer
          nullable: true
          minimum: 1
          maximum: 10
        type:
          type: string
          enum: [easy, tempo, intervals, long_run, race, fartlek, other]
        notes:
          type: string
        segments:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutSegment"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateWorkoutRequest:
      type: object
      required: [date, type]
      properties:
        date:
          type: string
          format: date
        distance_km:
          type: number
          format: float
        duration_seconds:
          type: integer
        avg_pace:
          type: string
        calories:
          type: integer
        avg_heart_rate:
          type: integer
        feeling:
          type: integer
          minimum: 1
          maximum: 10
        type:
          type: string
          enum: [easy, tempo, intervals, long_run, race, fartlek, other]
        notes:
          type: string
        segments:
          type: array
          items:
            $ref: "#/components/schemas/SegmentRequest"
    AssignedWorkout:
      type: object
      properties:
        id:
          type: integer
          format: int64
        coach_id:
          type: integer
          format: int64
        student_id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        type:
          type: string
        distance_km:
          type: number
          format: float
        duration_seconds:
          type: integer
        notes:
          type: string
        expected_fields:
          type: array
          items:
            type: string
        result_time_seconds:
          type: integer
          nullable: true
        result_distance_km:
          type: number
          format: float
          nullable: true
        result_heart_rate:
          type: integer
          nullable: true
        result_feeling:
          type: integer
          nullable: true
          minimum: 1
          maximum: 10
        image_file_id:
          type: integer
          format: int64
          nullable: true
        image_url:
          type: string
        status:
          type: string
          enum: [pending, completed, skipped]
        due_date:
          type: string
          format: date
        student_name:
          type: string
        coach_name:
          type: string
        segments:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutSegment"
        unread_message_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateAssignedWorkoutRequest:
      type: object
      required: [student_id, title, due_date]
      properties:
        student_id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        type:
          type: string
        distance_km:
          type: number
          format: float
        duration_seconds:
          type: integer
        notes:
          type: string
        expected_fields:
          type: array
          items:
            type: string
        due_date:
          type: string
          format: date
        segments:
          type: array
          items:
            $ref: "#/components/schemas/SegmentRequest"
...
    CoachStudentInfo:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
        avatar_url:
          type: string
...
    AdminUser:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        name:
          type: string
        avatar_url:
          type: string
        is_coach:
          type: boolean
        is_admin:
          type: boolean
        created_at:
          type: string
          format: date-time